thymyleaf

+++ src/main/kotlin/com/z/reactive/reactive/ReactiveApplication.kt
@@ -14,8 +14,10 @@ import org.springframework.data.mongodb.core.ReactiveMongoTemplate
 import org.springframework.data.mongodb.core.createCollection
 import reactor.core.publisher.Flux
 import reactor.core.publisher.toMono
+import springfox.documentation.swagger2.annotations.EnableSwagger2WebFlux
 
 @SpringBootApplication
+@EnableSwagger2WebFlux
 class ReactiveApplication(val reactiveMongoTemplate: ReactiveMongoTemplate, val registeredTemperatureDAO: RegisteredTemperatureDAO):ApplicationRunner {
     //on boot
     override fun run(args: ApplicationArguments?) {
@@ -23,18 +25,26 @@ class ReactiveApplication(val reactiveMongoTemplate: ReactiveMongoTemplate, val 
     }
 
     private fun reloadCappedCollection(){
-        reactiveMongoTemplate.dropCollection(RegisteredTemperature::class.java)
-                .then(
-                        reactiveMongoTemplate.createCollection(RegisteredTemperature::class.java, CollectionOptions.empty().capped().size(9128)) //equivalente a db.runCommand({ convertToCapped: 'registeredTemperature', size: 9128 })
-                                .thenMany(
-                                        registeredTemperatureDAO.saveAll(Flux.just(
-                                                RegisteredTemperature(id = null,device = "arduino",temperature = (0.00..100.00).random()),
-                                                RegisteredTemperature(id = null,device = "arduino",temperature = (0.00..100.00).random()),
-                                                RegisteredTemperature(id = null,device = "raspberry",temperature = (0.00..100.00).random()),
-                                                RegisteredTemperature(id = null,device = "raspberry",temperature = (0.00..100.00).random())
-                                        )))
-                                .toMono()
-                ).subscribe()
+        registeredTemperatureDAO.count().subscribe {
+            if(it==0L){
+                reactiveMongoTemplate.dropCollection(RegisteredTemperature::class.java)
+                        .then(
+                                reactiveMongoTemplate.createCollection(RegisteredTemperature::class.java, CollectionOptions.empty().capped().size(9128)) //equivalente a db.runCommand({ convertToCapped: 'registeredTemperature', size: 9128 })
+                                        .thenMany(
+                                                registeredTemperatureDAO.saveAll(Flux.just(
+                                                        RegisteredTemperature(id = null,device = "arduino",temperature = (0.00..100.00).random()),
+                                                        RegisteredTemperature(id = null,device = "arduino",temperature = (0.00..100.00).random()),
+                                                        RegisteredTemperature(id = null,device = "arduino",temperature = (0.00..100.00).random()),
+                                                        RegisteredTemperature(id = null,device = "arduino",temperature = (0.00..100.00).random()),
+                                                        RegisteredTemperature(id = null,device = "arduino",temperature = (0.00..100.00).random()),
+                                                        RegisteredTemperature(id = null,device = "raspberry",temperature = (0.00..100.00).random()),
+                                                        RegisteredTemperature(id = null,device = "raspberry",temperature = (0.00..100.00).random())
+                                                )))
+                                        .toMono()
+                        ).subscribe()
+            }
+        }
+
     }
 }
